Create README.md
To use this Payment Solutions Data Science project, please follow the steps below:

Ensure that you have Jupyter Notebook installed on your system. Jupyter Notebook allows you to run and interact with the code in this project.

Download the project files, including the Jupyter Notebook file (.ipynb) and the data file (data.csv).

Place both the Jupyter Notebook file and the data.csv file in the same folder on your local machine.

Open Jupyter Notebook and navigate to the folder where you saved the project files.

Click on the Jupyter Notebook file to open it.

Run each cell in the Jupyter Notebook sequentially to execute the code and see the results.